Playlist maker 0.8 Частично выстроена архитектура плеера и поиска
