--- app/src/main/java/com/practicum/playlistmaker/Data/retrofitSearch/SearchResponse.kt
@@ -1,8 +1,0 @@
-package com.practicum.playlistmaker.Data.retrofitSearch
-
-import com.practicum.playlistmaker.Data.Track
-
-data class SearchResponse(val results: ArrayList<Track>)
-
-
-
