--- app/src/main/java/com/practicum/playlistmaker/UI/MediaActivity.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmaker.UI
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.practicum.playlistmaker.R
-
-class MediaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_media)
-    }
-}
