--- app/src/main/java/com/practicum/playlistmaker/Utils/DateUtils.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.Utils
-
-import java.text.SimpleDateFormat
-import java.util.Locale
-
-object DateUtils {
-
-    fun changeDateFormat(date: String?) : String{
-        return SimpleDateFormat("mm:ss", Locale.getDefault()).format(date?.toInt())
-    }
-}
