+++ app/src/main/java/com/practicum/playlistmaker/data/dto/TracksSearchResponse.kt
@@ -1,0 +1,12 @@
+package com.practicum.playlistmaker.data.dto
+
+import com.practicum.playlistmaker.domain.models.Track
+
+data class TracksSearchResponse(
+    val results: ArrayList<Track>,
+//    val searchType: String,
+//    val expression: String
+) : Response()
+
+
+
