+++ app/src/main/java/com/practicum/playlistmaker/data/network/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.data.network
+
+import com.practicum.playlistmaker.data.dto.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any) : Response
+}
