+++ app/src/main/java/com/practicum/playlistmaker/data/network/TrackRepositoryImpl.kt
@@ -1,0 +1,31 @@
+package com.practicum.playlistmaker.data.network
+
+import com.practicum.playlistmaker.data.dto.TracksSearchRequest
+import com.practicum.playlistmaker.data.dto.TracksSearchResponse
+import com.practicum.playlistmaker.domain.models.Track
+import com.practicum.playlistmaker.domain.repository.TrackRepository
+
+class TrackRepositoryImpl(private val networkClient: NetworkClient) : TrackRepository {
+
+    override fun searchTracks(expression: String): List<Track> {
+        val response = networkClient.doRequest(TracksSearchRequest(expression))
+        if (response.resultCode == 200) {
+            return (response as TracksSearchResponse).results.map {
+                Track(
+                    it.trackId,
+                    it.trackName,
+                    it.artistName,
+                    it.trackTimeMillis,
+                    it.artworkUrl100,
+                    it.country,
+                    it.primaryGenreName,
+                    it.collectionName,
+                    it.previewUrl,
+                    it.releaseDate
+                ) }
+        } else {
+            return emptyList()
+        }
+    }
+
+}
