+++ app/src/main/java/com/practicum/playlistmaker/data/repository/utils/DateUtils.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.data.repository.utils
+
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+object DateUtils {
+
+    fun changeDateFormat(date: String?) : String{
+        return SimpleDateFormat("mm:ss", Locale.getDefault()).format(date?.toInt())
+    }
+}
