+++ app/src/main/java/com/practicum/playlistmaker/domain/repository/TrackRepository.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.domain.repository
+
+import com.practicum.playlistmaker.domain.models.Track
+
+interface TrackRepository {
+    fun searchTracks(expression: String) : List<Track>
+}
