+++ app/src/main/java/com/practicum/playlistmaker/domain/repository/player/playerManger/PlayerManager.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.domain.repository.player.playerManger
+
+import com.practicum.playlistmaker.domain.models.Track
+
+interface PlayerManager {
+
+    fun startPlayer()
+    fun pausePlayer()
+    fun preparePlayerForStart(track: Track?)
+    fun playbackControl()
+}
