+++ app/src/main/java/com/practicum/playlistmaker/domain/usecase/TracksInteractor.kt
@@ -1,0 +1,13 @@
+package com.practicum.playlistmaker.domain.usecase
+
+import com.practicum.playlistmaker.domain.models.Track
+
+interface TracksInteractor {
+
+    fun searchTracks(expression: String, consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(foundTracks: List<Track>)
+    }
+
+}
