+++ app/src/main/java/com/practicum/playlistmaker/presentation/ui/MainActivity.kt
@@ -1,0 +1,35 @@
+package com.practicum.playlistmaker.presentation.ui
+
+import android.annotation.SuppressLint
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.Button
+import com.practicum.playlistmaker.R
+
+class MainActivity : AppCompatActivity() {
+    @SuppressLint("MissingInflatedId")
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        val search = findViewById<Button>(R.id.search)
+        val media = findViewById<Button>(R.id.media)
+        val settings = findViewById<Button>(R.id.settings)
+
+        search.setOnClickListener {
+            val searchIntent = Intent(this, SearchActivity::class.java)
+            startActivity(searchIntent)
+        }
+
+        media.setOnClickListener {
+            val mediaIntent = Intent(this, MediaActivity::class.java)
+            startActivity(mediaIntent)
+        }
+
+        settings.setOnClickListener {
+            val settingsIntent = Intent(this, SettingsActivity::class.java)
+            startActivity(settingsIntent)
+        }
+    }
+
+}
