+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.MainActivity"
+    tools:context=".presentation.ui.MainActivity"
     style="@style/main_menu_background"
     android:paddingTop="14dp"
     android:paddingEnd="15dp"
